Update patch set 1
Patch Set 1:

yes, not ready yet - the actual "underrun" part is yet to be completed. i will do tests at the office with sip clients together with alex thsi week then..

the only important thing here is the point *when* and *how* contracts get locked, and the demarcation of the transaction - it covers *all* statements against billingdb. just saying, because this differs in libswrate (and it already did before this one).

Patch-set: 1
Label: Code-Review=0