MT#15463 work without provisioning db connection
provisioning db connectivity was added to ratomat
recently. to bring flexibility with the order of
how packages are upgraded, this fix allows ratomat
to run without provisioning db connectifity, i.e.
if ngcp-update-cfg-schema was not performed before.

note that it will have legacy behaviour in that case,
i.e. it will not lock subscribers upon balance
underrun.

+warning messages added

Change-Id: I50b1b8147a87e10806e594a994250bb173f5a584