Update patch set 2
Patch Set 2:

ok, it came from the original constructs to maintain the branch structure. i never tested



so the invocations of rollback() could be reduced to two - one in the inner catch (within the batch loop) and the outer one (within the main loop). a point is it's preferred to log the specific situation before rolling back.

Patch-set: 2