Update patch set 2
Patch Set 2:

crap editing ...

ok, it came from the original constructs to maintain the branch structure. i never tested

if ($DBI::err == 2006) {
} elsif ($DBI::err == 1213) {

so i left those part as is.

but agree, the invocations of rollback() could be reduced to two - one in the inner catch (within the batch loop)
and the outer one (within the main loop). a point to consider is that it's preferred to log the specific
situation message before rolling back.

Patch-set: 2
Label: Code-Review=0