TT#18115 Call start-stop-daemon correctly
We should pass the interpreter executable as --exec and the actual
script to run as --startas, as documented in the man page. Otherwise
start-stop-daemon will not be able to match the executable currently
running, and will consider that the daemon is *not* running. Use --name
to match on the process name.

Change-Id: Icb7a608aaf2dd3d54083f77ed6df67a7dbecdb8d