Update patch set 3
Patch Set 3:

well the print in there was only meant for viewing (and eventually capturing) when running the rate-o-mat testsuite.

since the print messages turn out to be the delayed ones, i thought to omit them to get rid of the duplication (which is needed anyways).

maybe the performance in alex' test runs show better performances without the prints. if that turns out true, there can be decision wether to globally disaable STDOUT capturing..?

Patch-set: 3
Label: Code-Review=0