TT#38400 disable updating the subscriber prepaid pref
we drop synchronizing the "prepaid" subscriber pref
to the value of prepaid flag of the profile (contract-level).

this removes the discrepancy with scheduling profiles
and, more important, allows to change a billing_profile's
prepaid setting. which we recently found several customers
are doing on regular basis.

every reviewer aware of a remaining use of the "prepaid"
subscriber pref, pls ping me.

this synchronisation part is also present in swrate, where it
can be disabled too.

Change-Id: Ia393b6d95ad4b4f7d7380cf491b1b150af83497f