TT#60950 Switch rate-o-mat docker to buster
Install mariadb-client instead of mysql-client, otherwise:
> E: Package 'mysql-client' has no installation candidate

Switch dependency from perl-modules to perl, as perl-modules
is not a public interface and we should not depends on
it directly (c) Guillem Jover.

P.S. also sorted list of modules alphabetically and remove unnecessary
docker layers by calling all apt-get in on "RUN".

Change-Id: Iea6ca4096a455dd91f76d6cc4f0f432334de3d9b