TT#62908 populate period cost sums and lock subscribers
after a cdr got rated and the call costs are calcualted,
rate-o-mat now populates 2 records in the
accounting.cdr_period_costs table for each source customer ID:

 - day record: accumulated cdr costs of the day of the
   call end time
 - month record: accumulated cdr costs of the month
   of the call end time

this works efficiently using upsert stmts (INSERT .. ON
DUPLICATE KEY ..) and therefor relies on the table's primary
key.

at the same time it checks if the sums exceeds either of
profile or contract fraud limits (if any). if either the
day or month fraud limit is exceeded and has has a fraud
lock level set, it will instantly apply that lock level to
all the customer ID's subscribers.

proper support for re-rating cdrs will follow in a separate
step.

Change-Id: I5a4f6139893c7ef01f343e75df8848342fab23e1