TT#66150 TT#42907 Fix b7c0308e: send systemd READY=1 when maintenance mode is enabled
The rate-o-mat is doing nothing when maintenance mode is enabled,
it has been implemented in TT#16754 but later after migration to
systemd and after migration to unit 'type=notify' we forget to send READY=1.
Fixing it here, as it cause 90 seconds delay on ngcpcfg apply on
active node if maintenance mode is enabled.

Change-Id: Ie0489635b4b9dc29e867517d4ae72417466d2e59