TT#88600 charge LNP number by type
so far, a fee with destination "lnp:<lnp_provider_id>" allows to charge
a call of a ported number by its lnp_provider_id. the problem with this
is that the lnp_provider_id changes each time you purge&upload new lnp
numbers.

this change adds supports for fee destinations such as
"lnpnumbertype:<type>" to match the lnp number's 'type' field. it is a
free text field, that can be set individually per lnp number.

Change-Id: Ic791ffc8d2a3ed63d5bf4d4a0812d8ac01529925