MT#56123 swap nodeA<->nodeB ID modulus/offsets
when repl is stuck (for whatever reason), each rate-o-mat
instance basically should proceed processing it's local
CDRs.

we observed it stalls though in that case.

the reason is because currently nodeA rate-o-mat is doing
even cdr IDs and nodeB is doing odd cdr IDs.

it has to be vice versa for the current offset configuration
in my.cnf. we might want to read the offset configuration
directly from the .cnf in a next step.

Change-Id: I0d6bafbd33f0d3370393a8cf1a763a7963bc017f